add user authentication

+++ app/Http/Controllers/TasksController.php
@@ -8,16 +8,22 @@ use App\Models\Task;
 
 class TasksController extends Controller
 {
-    // getでtasks/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        // タスク一覧を取得
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザのタスク一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
 
-        // タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        // dashboardビューでそれらを表示
+        return view('dashboard', $data);
     }
 
     // getでtasks/createにアクセスされた場合の「新規登録画面表示処理」
@@ -40,11 +46,11 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
 
-        // タスクを作成
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        // 認証済みユーザ（閲覧者）のタスクとして作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -53,13 +59,19 @@ class TasksController extends Controller
     // getでtasks/(任意のid)にアクセスされた場合の「取得表示(詳細表示)処理」
     public function show(string $id)
     {
-        // idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
+        // idの値で投稿を検索して取得
+        $task = \App\Models\Task::findOrFail($id);
 
-        // タスク詳細ビューをそれで表示
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        // 認証済みユーザ（閲覧者）がそのタスクの所有者である場合は,タスク詳細ビューを表示
+        if (\Auth::id() === $task->user_id) {
+            // タスク詳細ビューを表示
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }
+
+        // トップページへリダイレクトさせる
+        return redirect('/');
     }
 
     // getでtasks/(任意のid)/editにアクセスされた場合の「更新画面表示処理」
@@ -67,12 +79,17 @@ class TasksController extends Controller
     {
 
         // idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
+        $task = \App\Models\Task::findOrFail($id);
 
-        // タスク編集ビューでそれを表示
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        // // 認証済みユーザ（閲覧者）がそのタスクの所有者である場合は,タスク編集ビューでそれを表示
+            if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+            }
+
+            // トップページへリダイレクトさせる
+            return redirect('/');
     }
 
     // putまたはpatchでtasks/(任意のid)にアクセスされた場合の「更新処理」
@@ -86,7 +103,8 @@ class TasksController extends Controller
 
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        // タスクを更新
+
+        // 認証済みユーザ（閲覧者）のタスクとして更新（リクエストされた値をもとに作成）
         $task->content = $request->content;
         $task->status = $request->status;
         $task->save();
@@ -98,12 +116,16 @@ class TasksController extends Controller
     // deleteでtasks/(任意のid)にアクセスされた場合の「削除処理」
     public function destroy(string $id)
     {
-        // idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
+        // idの値で投稿を検索して取得
+        $task = \App\Models\Task::findOrFail($id);
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return redirect('/')->with('success', 'Delete Successful');
+        }
 
         // トップページへリダイレクトさせる
-        return redirect('/');
+        return redirect('/')->with('Delete Failed'); 
     }
 }
